Completed the user login flow and update the contacts api's. Also, added the middleware called auth to valide user token associated with it.

+++ src/controller/contactController.js
@@ -3,16 +3,16 @@ const contactModel = require("../models/contactModel");
 /*
     @des: Get all contacts
     @route: /api/contacts
-    @access: public 
+    @access: private 
 */
 const getAllContacts = asyncHandler(async (req, res) => {
-    const contactList = await contactModel.find();
+    const contactList = await contactModel.find({userId: req.user.id});
     res.status(200).json(contactList);
 });
 /*
     @des: Get single contact
     @route: /api/contacts:id
-    @access: public
+    @access: private
 */
 const getContact = asyncHandler(async (req, res) => {
     const singleContact = await contactModel.findById(req.params.id);
@@ -25,7 +25,7 @@ const getContact = asyncHandler(async (req, res) => {
 /*
     @des : Create contact
     @route: /api/contacts
-    @access: public 
+    @access: private 
 */
 const createContact = asyncHandler(async (req, res) => {
     const { name, email, phone } = req.body;
@@ -41,14 +41,15 @@ const createContact = asyncHandler(async (req, res) => {
     const newContact = await contactModel.create({
         name,
         email,
-        phone
+        phone,
+        userId: req.user.id
     });
     res.status(201).json(newContact);
 });
 /*
     @des: Update contact
     @route: /api/contacts:id
-    @access: public 
+    @access: private 
 */
 const updateContact = asyncHandler(async (req, res) => {
     const contact = await contactModel.findById(req.params.id);
@@ -56,10 +57,10 @@ const updateContact = asyncHandler(async (req, res) => {
         res.status(404);
         throw new Error("Contact not found.");
     }
-    const checkUser = await contactModel.find({ "email": req.body.email });
-    if (checkUser.length > 0) {
-        res.status(400);
-        throw new Error("Email already linked with an account.");
+    
+    if(contact.userId.toString() !== req.user.id) {
+        res.status(403);
+        throw new Error("User not authorized.");
     }
     //updating 
     const updatedContact = await contactModel.findByIdAndUpdate(req.params.id, req.body, { new: true });
@@ -75,6 +76,10 @@ const deleteContact = asyncHandler(async (req, res) => {
     if (!contact) {
         res.status(404);
         throw new Error("Contact not found.");
+    }
+    if(contact.userId.toString() !== req.user.id) {
+        res.status(403);
+        throw new Error("User not authorized.");
     }
     await contactModel.findByIdAndRemove(req.params.id);
     res.status(200).json(contact);
